+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Releases
 
 ## Unreleased
+#### Features
+  - The load log file can now be configured to include relations (#94)
+
 #### Bugfixes
   - Fixed an access violation which could occur when loading from CSV or SQL auxilliary tables (#92)